move some typed functions to types.js
